Raise Exception if spaces in Game or Game agent names
If the user attempts to generate a game plugin or game agent plugin they are prompted to enter a name that contains no spaces. This patch now raises an exception if there is a space found in the string entered by the user while attempting to generate the game/game agent plugin.